Changes to wipId

+++ KY-MES.Domain/V1/DTOs/OutputModels/AddDefectRequestModel.cs
@@ -9,14 +9,14 @@ namespace KY_MES.Domain.V1.DTOs.OutputModels
 {
     public class AddDefectRequestModel
     {
-        public int? wipId { get; set; }
+        public long? wipId { get; set; }
         public List<Defect> defects { get; set; }
         public bool? hasValidNumericField { get; set; }
         public List<PanelDefect> panelDefects { get; set; }
     }
     public class PanelDefect
     {
-        public int? wipId { get; set; }
+        public long? wipId { get; set; }
         public List<Defect> defects { get; set; }
         public bool? hasValidNumericField { get; set; }
     }
